refactor(errors) Move location extraction into a utilty module
