+++ packages/cicero-core/index.js
@@ -19,6 +19,7 @@
  * @module cicero-core
  */
 
+module.exports.ErrorUtil = require('./lib/errorutil');
 module.exports.Clause = require('./lib/clause');
 module.exports.Template = require('./lib/template');
 module.exports.TemplateLibrary = require('./lib/templatelibrary');
