+++ packages/cicero-core/test/module.js
@@ -27,8 +27,10 @@ describe('Module', () => {
     describe('#exports', () => {
 
         it('should export classes', () => {
+            index.ErrorUtil.should.not.be.null;
             index.Clause.should.not.be.null;
             index.Template.should.not.be.null;
+            index.TemplateLibrary.should.not.be.null;
             index.version.should.not.be.null;
         });
     });
